Split tests. Improve aspect.
