+++ tracer/test/test-applications/integrations/Samples.InstrumentedTests/Vulnerabilities/SSRF/RestClientTests.cs
@@ -1,0 +1,53 @@
+using Xunit;
+using RestSharp;
+using System.Threading;
+
+namespace Samples.InstrumentedTests.Iast.Vulnerabilities.SSRF;
+
+public class RestClientTests : SSRFTests
+{
+    [Fact]
+    public void GivenARestSharpClient_WhenExecute_Vulnerable()
+    {
+        var client = new RestClient(taintedUrlValue);
+        var request = new RestRequest("", Method.Get);
+        ExecuteAction(() => client.Execute(request));
+        AssertVulnerable("SSRF", sourceType: sourceType);
+    }
+
+    [Fact]
+    public void GivenARestSharpClient_WhenExecute_Vulnerable2()
+    {
+        var client = new RestClient(taintedUrlValue);
+        var request = new RestRequest("", Method.Get);
+        ExecuteAction(() => client.ExecuteAsync(request, CancellationToken.None));
+        AssertVulnerable("SSRF", sourceType: sourceType);
+    }
+
+    [Fact]
+    public void GivenARestSharpClient_WhenExecute_Vulnerable3()
+    {
+        var client = new RestClient(taintedUrlValue);
+        var request = new RestRequest("", Method.Get);
+        ExecuteAction(() => client.Delete(request));
+        AssertVulnerable("SSRF", sourceType: sourceType);
+    }
+
+    [Fact]
+    public void GivenARestSharpClient_WhenExecute_Vulnerable4()
+    {
+        var client = new RestClient(taintedUrlValue);
+        var request = new RestRequest("", Method.Get);
+        ExecuteAction(() => client.Post(request));
+        AssertVulnerable("SSRF", sourceType: sourceType);
+    }
+
+    [Fact]
+    public void GivenARestSharpClient_WhenExecute_Vulnerable5()
+    {
+        var client = new RestClient(taintedUrlValue);
+        var request = new RestRequest("", Method.Get);
+        ExecuteAction(() => client.Get(request));
+        AssertVulnerable("SSRF", sourceType: sourceType);
+    }
+}
