+++ tracer/test/test-applications/integrations/Samples.InstrumentedTests/Vulnerabilities/SSRF/UriTests.cs
@@ -1,14 +1,11 @@
 using System;
-using System.Collections.Specialized;
-using System.Net;
-using RestSharp;
-using Moq;
-using System.Net.Http;
-using System.Threading;
 using Xunit;
 using FluentAssertions;
+#if NET6_0_OR_GREATER
+using System.Net.Http;
+#endif
 
-namespace Samples.InstrumentedTests.Iast.Vulnerabilities;
+namespace Samples.InstrumentedTests.Iast.Vulnerabilities.SSRF;
 
 public class UriTests : SSRFTests
 {
@@ -96,14 +93,14 @@ public class UriTests : SSRFTests
     [Fact]
     public void GivenAURI_WhenCreateFromtainted_IsTainted13()
     {
-        Uri uri = new Uri(new Uri(taintedUrlValue), new Uri(notTaintedHost));
+        Uri uri = new Uri(new Uri(taintedUrlValue), new Uri(notTaintedHost, UriKind.Relative));
         AssertTainted(uri.OriginalString);
     }
 
     [Fact]
     public void GivenAURI_WhenCreateFromtainted_IsTainted14()
     {
-        Uri uri = new Uri(new Uri(notTaintedValue), new Uri(taintedHost));
+        Uri uri = new Uri(new Uri(notTaintedValue), new Uri(taintedHost, UriKind.Relative));
         AssertTainted(uri.OriginalString);
     }
 
@@ -156,8 +153,8 @@ public class UriTests : SSRFTests
     [Fact]
     public void GivenAURI_WhenTryCreateFromtainted_IsTainted5()
     {
-        Uri uri;
-        bool result = Uri.TryCreate(new Uri(taintedUrlValue), new Uri("relative"), out uri);
+        Uri uri = new Uri(taintedUrlValue);
+        bool result = Uri.TryCreate(new Uri(taintedUrlValue), new Uri("relative", UriKind.Relative), out uri);
         result.Should().BeTrue();
         AssertTainted(uri.OriginalString);
     }
@@ -166,7 +163,7 @@ public class UriTests : SSRFTests
     public void GivenAURI_WhenTryCreateFromtainted_IsTainted6()
     {
         Uri uri;
-        bool result = Uri.TryCreate(new Uri(notTaintedValue), new Uri(taintedHost), out uri);
+        bool result = Uri.TryCreate(new Uri(notTaintedValue), new Uri(taintedHost, UriKind.Relative), out uri);
         result.Should().BeTrue();
         AssertTainted(uri.OriginalString);
     }
@@ -194,5 +191,48 @@ public class UriTests : SSRFTests
         AssertTaintedFormatWithOriginalCallCheck(
             Uri.EscapeDataString(taintedUrlValue),
             () => Uri.EscapeDataString(taintedUrlValue));
+    }
+
+    [Fact]
+    [Obsolete]
+    public void GivenAURI_WhenMakeRelativeFromtainted_IsTainted()
+    {
+        Uri uri = new Uri(taintedUrlValue2);
+        var uriRelative = uri.MakeRelative(new Uri(taintedUrlValue));
+        uriRelative.Should().NotBeNullOrWhiteSpace();
+        AssertTainted(uriRelative);
+    }
+
+    [Fact]
+    public void GivenAURI_WhenMakeRelativeFromtainted_IsTainted2()
+    {
+        Uri uri = new Uri(taintedUrlValue2);
+        var uriRelative = uri.MakeRelativeUri(new Uri(taintedUrlValue));
+        uriRelative.Should().NotBeNull();
+        AssertTainted(uriRelative.OriginalString);
+    }
+
+    [Fact]
+    public void GivenAUriTainted_WhengetProperties_tainted()
+    {
+        var uri = new Uri(taintedUrlValue);
+        AssertTainted(uri.AbsoluteUri);
+        AssertTainted(uri.ToString());
+        AssertTainted(uri.AbsolutePath);
+        AssertTainted(uri.Scheme);
+        AssertTainted(uri.Query);
+        AssertTainted(uri.Authority);
+        AssertTainted(uri.Host);
+        AssertTainted(uri.LocalPath);
+        AssertTainted(uri.OriginalString);
+        AssertTainted(uri.PathAndQuery);
+        AssertNotTainted(uri.UserInfo); // user info is empty
+    }
+
+    [Fact]
+    public void GivenAUriTainted_WhengetPropertiesMultipleTimes_RangesAreNotDuplicated()
+    {
+        var uri = new Uri(AddTainted("http://localhost/tainted?with=qs").ToString());
+        AssertTainted(uri.Query);
     }
 }
